retester: interfacing backoffCache
